Adding new files in refactor
