+++ src/MeshUtils.jl
@@ -1,0 +1,27 @@
+# mesh tools
+
+meshgrid(v::AbstractVector) = meshgrid(v, v)
+
+function meshgrid(vx::AbstractVector{T}, vy::AbstractVector{T}) where T
+    m, n = length(vy), length(vx)
+    vx = reshape(vx, 1, n)
+    vy = reshape(vy, m, 1)
+    (repmat(vx, m, 1), repmat(vy, 1, n))
+end
+
+
+# Construct mesh for quick reconstruction
+function buildmesh!(dc::DepthCamera)
+  H, W = dc.shape
+  xs,ys = collect(1:W), collect(1:H)
+  fxinv = 1.0 / dc.K[1,1];
+  fyinv = 1.0 / dc.K[2,2];
+
+  xs = (xs-dc.K[1,3]) * fxinv
+  xs = xs[1:dc.skip:end]
+  ys = (ys-dc.K[2,3]) * fyinv
+  ys = ys[1:dc.skip:end]
+
+  dc.xs, dc.ys = meshgrid(xs, ys);
+  nothing
+end
